refactor: split tests up; do not parse when injecting attr
